@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-	<TargetFramework>net6.0</TargetFramework>
+	<TargetFramework>net7.0</TargetFramework>
 	<Nullable>enable</Nullable>
 	<ImplicitUsings>enable</ImplicitUsings>
 	<UserSecretsId>CookTo.Server-5B296B08-D554-4243-9E88-AB1917E222E9</UserSecretsId>
@@ -9,17 +9,20 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <None Remove="Endpoints\RecipeSummaryEndpointsExtensions.cs~RF41abb4.TMP" />
+    <Compile Remove="NewFolder\**" />
+    <Content Remove="NewFolder\**" />
+    <EmbeddedResource Remove="NewFolder\**" />
+    <None Remove="NewFolder\**" />
   </ItemGroup>
 
   <ItemGroup>
 	  <PackageReference Include="AutoMapper" Version="11.0.1" />
-	  <PackageReference Include="AutoMapper.Collection" Version="7.0.1" />
+	  <PackageReference Include="AutoMapper.Collection" Version="8.0.0" />
 	  <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="11.0.0" />
-	  <PackageReference Include="FluentValidation.AspNetCore" Version="10.3.6" />
-	<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Server" Version="6.0.2" />
-	  <PackageReference Include="MongoDB.Driver" Version="2.14.1" />
-	  <PackageReference Include="SixLabors.ImageSharp" Version="2.0.0" />
+	  <PackageReference Include="FluentValidation.AspNetCore" Version="11.0.2" />
+	<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Server" Version="7.0.0-preview.4.22251.1" />
+	  <PackageReference Include="MongoDB.Driver" Version="2.15.1" />
+	  <PackageReference Include="SixLabors.ImageSharp" Version="2.1.2" />
 
 	  
   </ItemGroup>
@@ -30,10 +33,10 @@
   </ItemGroup>
 
   <ItemGroup>
-	<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="6.0.2" NoWarn="NU1605" />
-	<PackageReference Include="Microsoft.AspNetCore.Authentication.OpenIdConnect" Version="6.0.2" NoWarn="NU1605" />
-	<PackageReference Include="Microsoft.Identity.Web" Version="1.22.3" />
-	<PackageReference Include="Microsoft.Identity.Web.UI" Version="1.22.3" />
+	<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="7.0.0-preview.4.22251.1" NoWarn="NU1605" />
+	<PackageReference Include="Microsoft.AspNetCore.Authentication.OpenIdConnect" Version="7.0.0-preview.4.22251.1" NoWarn="NU1605" />
+	<PackageReference Include="Microsoft.Identity.Web" Version="1.25.0" />
+	<PackageReference Include="Microsoft.Identity.Web.UI" Version="1.25.0" />
   </ItemGroup>
 
   <ItemGroup>
