@@ -1,4 +1,4 @@
-using CookTo.Shared.Features.ManageRecipes;
+using CookTo.Shared.Modules.ManageRecipes;
 using Microsoft.AspNetCore.Components.Forms;
 using Newtonsoft.Json;
 
@@ -12,19 +12,21 @@ public class UploadImageManager : IUploadImageManager
 
     public UploadImageManager(IHttpClientFactory factory) => _factory = factory;
 
-    public async Task<IBrowserFile> GetImage(string recipeId)
+#pragma warning disable CS8613 // Nullability of reference types in return type doesn't match implicitly implemented member.
+    public async Task<IBrowserFile?> GetImage(string recipeId)
+#pragma warning restore CS8613 // Nullability of reference types in return type doesn't match implicitly implemented member.
     {
         try
         {
-            var httpClient = HttpClientFactoryHelper.CreateClient(_factory, HttpClientType.Anon);
+            var httpClient = HttpNamedClientFactoryHelper.CreateClient(_factory, HttpClientType.Anon);
 
             var result = await httpClient.GetAsync($"{_url}/{recipeId}", new CancellationToken());
             result.EnsureSuccessStatusCode();
             var content = await result.Content.ReadAsStringAsync();
-            var response = JsonConvert.DeserializeObject<IBrowserFile>(content);
-            return response;
-        }
-        catch (HttpRequestException)
+            var deserializedObject = JsonConvert.DeserializeObject<IBrowserFile>(content);
+            return deserializedObject;       
+            
+        } catch(HttpRequestException)
         {
             return default!;
         }
@@ -34,18 +36,17 @@ public class UploadImageManager : IUploadImageManager
     {
         try
         {
-            MemoryStream ms = new MemoryStream();
+            MemoryStream ms = new();
             await file.OpenReadStream(file.Size).CopyToAsync(ms);
-            var dto = new ImageUploadDto() { RecipeId = recipeId, Image = ms.ToArray() };
+            var dto = new ImageUpload() { RecipeId = recipeId, Image = ms.ToArray() };
 
-            var httpClient = HttpClientFactoryHelper.CreateClient(_factory, HttpClientType.Secure);
+            var httpClient = HttpNamedClientFactoryHelper.CreateClient(_factory, HttpClientType.Secure);
 
             var result = await httpClient.PostAsJsonAsync(_url, dto, new CancellationToken());
             result.EnsureSuccessStatusCode();
             var content = await result.Content.ReadAsStringAsync();
             return content;
-        }
-        catch (Exception)
+        } catch(Exception)
         {
             throw;
         }
