@@ -1,7 +1,7 @@
 ﻿@inject AppState AppState
 
 <CascadingAuthenticationState>
-
+    <CascadingBlazoredModal>
     <Router AppAssembly="@typeof(App).Assembly">
         <Found Context="routeData">
             <AuthorizeRouteView RouteData="@routeData" DefaultLayout="@typeof(MainLayout)">
@@ -25,6 +25,7 @@
             </LayoutView>
         </NotFound>
     </Router>  
+    </CascadingBlazoredModal>
 </CascadingAuthenticationState>
 
 @code {
