@@ -1,5 +1,4 @@
-﻿
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 
 namespace CookTo.Client.Managers;
 
@@ -18,41 +17,66 @@ public abstract class BaseManager<T> : IBaseManager<T> where T : class
     {
         try
         {
-            var httpClient = HttpClientFactoryHelper.CreateClient(_httpClientFactory, HttpClientType.Anon);
+            var httpClient = HttpNamedClientFactoryHelper.CreateClient(_httpClientFactory, HttpClientType.Anon);
             var result = await httpClient.GetAsync(_url, new CancellationToken());
             var content = await result.Content.ReadAsStringAsync();
-            return JsonConvert.DeserializeObject<List<T>>(content);
+            var deserializedObject = JsonConvert.DeserializeObject<IList<T>>(content);
+            if (deserializedObject != null)
+            {
+                return deserializedObject;
+            }
+            else
+            {
+                return new List<T>();
+            }
         } catch(HttpRequestException)
         {
             return default!;
         }
     }
 
-    public async Task<T> GetById(string id)
+    public async Task<T?> GetById(string id)
     {
         try
         {
-            var httpClient = HttpClientFactoryHelper.CreateClient(_httpClientFactory, HttpClientType.Anon);
+            var httpClient = HttpNamedClientFactoryHelper.CreateClient(_httpClientFactory, HttpClientType.Anon);
             var result = await httpClient.GetAsync($"{_url}/{id}", new CancellationToken());
             result.EnsureSuccessStatusCode();
             var content = await result.Content.ReadAsStringAsync();
-            return JsonConvert.DeserializeObject<T>(content);
+            var deserializedObject = JsonConvert.DeserializeObject<T>(content);
+            if (deserializedObject != null)
+            {
+                return deserializedObject;
+            }
+            else
+            {
+                return null;
+            }
         } catch(HttpRequestException)
         {
             return default!;
         }
     }
 
-    public async Task<T> Insert(T entity)
+    public async Task<T?> Insert(T entity)
     {
         try
         {
-            var httpClient = HttpClientFactoryHelper.CreateClient(_httpClientFactory, HttpClientType.Secure);
+            var httpClient = HttpNamedClientFactoryHelper.CreateClient(_httpClientFactory, HttpClientType.Secure);
 
             var result = await httpClient.PostAsJsonAsync(_url, entity, new CancellationToken());
             result.EnsureSuccessStatusCode();
             var content = await result.Content.ReadAsStringAsync();
-            return JsonConvert.DeserializeObject<T>(content);
+           var deserializedObject = JsonConvert.DeserializeObject<T>(content);
+            if (deserializedObject != null)
+            {
+                return deserializedObject;
+            }
+            else
+            {
+                return null;
+            }
+        
         } catch(Exception)
         {
             throw;
@@ -63,7 +87,7 @@ public abstract class BaseManager<T> : IBaseManager<T> where T : class
     {
         try
         {
-            var httpClient = HttpClientFactoryHelper.CreateClient(_httpClientFactory, HttpClientType.Secure);
+            var httpClient = HttpNamedClientFactoryHelper.CreateClient(_httpClientFactory, HttpClientType.Secure);
             var result = await httpClient.PutAsJsonAsync(_url, entityToUpdate);
             return result.IsSuccessStatusCode;
         } catch(Exception)
