@@ -1,8 +1,8 @@
-using CookTo.Shared.Features.ManageCategory;
+using CookTo.Shared.Modules.ManageCategories;
 
 namespace CookTo.Client.Managers;
 
-public class CategoryManager : BaseManager<CategoryDto>
+public class CategoryManager : BaseManager<Category>
 {
     public CategoryManager(IHttpClientFactory factory) : base(factory, "category")
     {
