@@ -1,8 +1,8 @@
-using CookTo.Shared.Features.ManageCuisine;
+using CookTo.Shared.Modules.ManageCuisines;
 
 namespace CookTo.Client.Managers;
 
-public class CuisineManager : BaseManager<CuisineDto>
+public class CuisineManager : BaseManager<Cuisine>
 {
     public CuisineManager(IHttpClientFactory factory) : base(factory, "cuisine")
     {
