@@ -1,4 +1,4 @@
-using CookTo.Shared.Features.ManageRecipes;
+using CookTo.Shared.Modules.ManageRecipes;
 using Newtonsoft.Json;
 
 namespace CookTo.Client.Managers;
@@ -12,16 +12,23 @@ public class HighlightedRecipeManager : IHighlightedRecipeManager
 
     public HighlightedRecipeManager(IHttpClientFactory factory) { _factory = factory; }
 
-    public async Task<HighlightedRecipeDto> GetHighlighted()
+    public async Task<HighlightedRecipe?> GetHighlighted()
     {
         try
         {
-            var httpClient = HttpClientFactoryHelper.CreateClient(_factory, HttpClientType.Anon);
+            var httpClient = HttpNamedClientFactoryHelper.CreateClient(_factory, HttpClientType.Anon);
             var result = await httpClient.GetAsync(_url, new CancellationToken());
-            result.EnsureSuccessStatusCode();
+                result.EnsureSuccessStatusCode();
             var content = await result.Content.ReadAsStringAsync();
-            var response = JsonConvert.DeserializeObject<HighlightedRecipeDto>(content);
-            return response;
+            var deserializedObject = JsonConvert.DeserializeObject<HighlightedRecipe>(content);
+            if (deserializedObject != null)
+            {
+                return deserializedObject;
+            }
+            else
+            {
+                return null;
+            }
         } catch(Exception)
         {
             return default!;
