@@ -1,8 +1,8 @@
-using CookTo.Shared.Features.ManageIngredients;
+using CookTo.Shared.Modules.ManageIngredients;
 
 namespace CookTo.Client.Managers;
 
-public class IngredientManager : BaseManager<IngredientDto>
+public class IngredientManager : BaseManager<Ingredient>
 {
     public IngredientManager(IHttpClientFactory factory) : base(factory, "ingredient")
     {
