@@ -4,9 +4,9 @@ public interface IBaseManager<T>
 {
     Task<IList<T>> GetAll();
 
-    Task<T> GetById(string id);
+    Task<T?> GetById(string id);
 
-    Task<T> Insert(T entity);
+    Task<T?> Insert(T entity);
 
     Task<bool> Update(T entityToUpdate);
 }
