@@ -1,9 +1,8 @@
-using CookTo.Shared.Features.ManageRecipes;
-
+using CookTo.Shared.Modules.ManageRecipes;
 
 namespace CookTo.Client.Managers.Interfaces;
 
 public interface IHighlightedRecipeManager
 {
-    Task<HighlightedRecipeDto> GetHighlighted();
+    Task<HighlightedRecipe?> GetHighlighted();
 }
