@@ -1,11 +1,10 @@
-﻿
-using CookTo.Shared.Features.ManageRecipes;
+﻿using CookTo.Shared.Modules.ManageRecipes;
 
 namespace CookTo.Client.Managers.Interfaces;
 
 public interface IRecipeSummaryManager
 {
-    Task<IEnumerable<RecipeSummaryDto>> GetCount(int amount);
+    Task<IEnumerable<RecipeSummary>> GetCount(int amount);
 
-    Task<IEnumerable<RecipeSummaryDto>> GetByTerm(string id);
+    Task<IEnumerable<RecipeSummary>> GetByTerm(string id);
 }
