@@ -1,8 +1,8 @@
-using CookTo.Shared.Features.ManageRecipes;
+using CookTo.Shared.Modules.ManageRecipes;
 
 namespace CookTo.Client.Managers;
 
-public class RecipeManager : BaseManager<RecipeDto>
+public class RecipeManager : BaseManager<FullRecipe>
 {
     public RecipeManager(IHttpClientFactory factory) : base(factory, "recipe")
     {
