@@ -1,8 +1,8 @@
-using CookTo.Shared.Features.ManageUtensils;
+using CookTo.Shared.Modules.ManageUtensils;
 
 namespace CookTo.Client.Managers;
 
-public class UtensilManager : BaseManager<UtensilDto>
+public class UtensilManager : BaseManager<Utensil>
 {
     public UtensilManager(IHttpClientFactory factory) : base(factory, "utensil")
     {
