@@ -1,4 +1,5 @@
 using Blazored.LocalStorage;
+using Blazored.Modal;
 using CookTo.Client;
 using CookTo.Client.Managers;
 using CookTo.Client.State;
@@ -38,11 +39,10 @@ builder.Services.AddScoped<RecipeManager>();
 builder.Services.AddScoped<CuisineManager>();
 builder.Services.AddScoped<IHighlightedRecipeManager, HighlightedRecipeManager>();
 builder.Services.AddScoped<IRecipeSummaryManager, RecipeSummaryManager>();
-builder.Services.AddScoped<IFavoritesManager, FavoritesManager>();
 builder.Services.AddScoped<IUploadImageManager, UploadImageManager>();
 
 builder.Services.AddScoped<AppState>();
 
 builder.Services.AddBlazoredLocalStorage();
-
+builder.Services.AddBlazoredModal();
 await builder.Build().RunAsync();
