@@ -2,17 +2,17 @@
 
 <div class="card shadow" >
     
-        <img alt="Image of @RecipeSummary.RecipeTitle Recipe"    src="@(!string.IsNullOrWhiteSpace(RecipeSummary.ImageFileName) ? $"images/{RecipeSummary.ImageFileName}" : "https://via.placeholder.com/286x190.jpg?text=No+Image+For+Trail")" class="card-img-top ">
+        <img alt="Image of @RecipeSummary.Title Recipe"    src="@(!string.IsNullOrWhiteSpace(RecipeSummary.Image) ? $"images/{RecipeSummary.Image}" : "https://via.placeholder.com/286x190.jpg?text=No+Image+For+Recipe")" class="card-img-top ">
       <div class="pt-2 ps-2">
-         <h7 class="card-title">@RecipeSummary.RecipeTitle</h7>  
+         <h7 class="card-title">@RecipeSummary.Title</h7>  
        
         <div class="d-flex flex-row mb-1 justify-content-between">    
         <small> By: @RecipeSummary.AddedBy</small>
-      <small class="text-muted pe-1"><span class="bi bi-grid"></span>  @RecipeSummary.Category.Name</small>  
+      <small class="text-muted pe-1"><span class="bi bi-grid"></span>  @RecipeSummary.Category.Text</small>  
       
         </div>
          <div class="d-flex flex-row-reverse gap-1 p-2">
-        <button aria-label="viewRecipe" class="btn btn-primary" title="View" @onclick="@(async () => NavManager.NavigateTo($"/recipe/{RecipeSummary.RecipeId}"))">
+        <button aria-label="viewRecipe" class="btn btn-primary" title="View" @onclick="@(async () => NavManager.NavigateTo($"/recipe/{RecipeSummary.Id}"))">
             <i class="bi bi-file-earmark-text"></i>
         </button>
 
@@ -20,7 +20,7 @@
         <AuthorizeView>
             @if (RecipeSummary.AddedBy.Equals(context.User?.Identity?.Name, StringComparison.CurrentCultureIgnoreCase) || context.User.IsInRole("Administrator"))
             {
-                <button aria-Label="editRecipe" class="btn btn-outline-secondary float-right" title="Edit" @onclick="@(() => NavManager.NavigateTo($"/edit-recipe/{RecipeSummary.RecipeId}"))">
+                <button aria-Label="editRecipe" class="btn btn-outline-secondary float-right" title="Edit" @onclick="@(() => NavManager.NavigateTo($"/edit-recipe/{RecipeSummary.Id}"))">
                     <i class="bi bi-pencil"></i>
                 </button>
             }
@@ -31,6 +31,6 @@
 </div>
 
 @code {
-    [Parameter, EditorRequired] public RecipeSummaryDto RecipeSummary { get; set; } 
+    [Parameter, EditorRequired] public RecipeSummary RecipeSummary { get; set; } 
     
 }
