@@ -8,26 +8,32 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
+
 @using CookTo.Client
-@using CookTo.Client.Features.Auth
-@using CookTo.Client.Features.Home
-@using CookTo.Client.Features.Home.Shared
-@using CookTo.Client.Features.Layout
-@using CookTo.Client.Features.ManageRecipes.DisplaySections
-@using CookTo.Client.Features.ManageRecipes.FormParts
-@using CookTo.Client.Features.ManageRecipes.Shared
-@using CookTo.Client.Features.ManageRecipes.Typeaheads
+@using CookTo.Client.Modules.Auth
+@using CookTo.Client.Modules.Home
+@using CookTo.Client.Modules.Home.Shared
+@using CookTo.Client.Modules.Layout
+@using CookTo.Client.Modules.ManageRecipes.DisplaySections
+@using CookTo.Client.Modules.ManageRecipes.FormParts
+@using CookTo.Client.Modules.ManageRecipes.Shared
+@using CookTo.Client.Modules.ManageRecipes.TypeAheads
 @using CookTo.Client.Managers
 @using CookTo.Client.Managers.Interfaces
 @using CookTo.Client.Shared
 @using CookTo.Client.State
 @using CookTo.Client.Validation
+
+
 @using CookTo.Shared.Enums
-@using CookTo.Shared.Features.ManageCuisine
-@using CookTo.Shared.Features.ManageCategory
-@using CookTo.Shared.Features.ManageFavorites
-@using CookTo.Shared.Features.ManageIngredients
-@using CookTo.Shared.Features.ManageRecipes
-@using CookTo.Shared.Features.ManageUtensils
+@using CookTo.Shared.Modules.ManageRecipes
+@using CookTo.Shared.Modules.ManageIngredients
+@using CookTo.Shared.Modules.ManageCategories
+@using CookTo.Shared.Modules.ManageCuisines
+@using CookTo.Shared.Modules.ManageUtensils
+
+
 @using Blazored.FluentValidation
 @using Blazored.Typeahead
+@using Blazored.Modal
+@using Blazored.Modal.Services
